+++ ingenialink/canopen/dictionary.py
@@ -1,7 +1,7 @@
 from typing import Optional, List
 
 from ingenialink.dictionary import DictionaryV2, Interface
-from ingenialink.canopen.register import CanopenRegister, REG_DTYPE, REG_ACCESS
+from ingenialink.canopen.register import CanopenRegister, REG_DTYPE, REG_ACCESS, RegCyclicType
 
 import ingenialogger
 import xml.etree.ElementTree as ET
@@ -22,8 +22,8 @@ class CanopenDictionaryV2(DictionaryV2):
             identifier="MONITORING_DATA",
             idx=0x58B2,
             subidx=0x00,
-            cyclic="CONFIG",
-            dtype=REG_DTYPE.U16,
+            cyclic=RegCyclicType.CONFIG,
+            dtype=REG_DTYPE.BYTE_ARRAY_512,
             access=REG_ACCESS.RO,
             subnode=0,
         ),
@@ -31,8 +31,8 @@ class CanopenDictionaryV2(DictionaryV2):
             identifier="DISTURBANCE_DATA",
             idx=0x58B4,
             subidx=0x00,
-            cyclic="CONFIG",
-            dtype=REG_DTYPE.U16,
+            cyclic=RegCyclicType.CONFIG,
+            dtype=REG_DTYPE.BYTE_ARRAY_512,
             access=REG_ACCESS.WO,
             subnode=0,
         ),