+++ ingenialink/ethernet/dictionary.py
@@ -4,7 +4,7 @@ import xml.etree.ElementTree as ET
 import ingenialogger
 
 from ingenialink.dictionary import DictionaryV2, Interface
-from ingenialink.ethernet.register import EthernetRegister, REG_DTYPE, REG_ACCESS
+from ingenialink.ethernet.register import EthernetRegister, REG_DTYPE, REG_ACCESS, RegCyclicType
 
 logger = ingenialogger.get_logger(__name__)
 
@@ -23,8 +23,8 @@ class EthernetDictionaryV2(DictionaryV2):
             units="",
             subnode=0,
             address=0x00B2,
-            cyclic="CONFIG",
-            dtype=REG_DTYPE.U16,
+            cyclic=RegCyclicType.CONFIG,
+            dtype=REG_DTYPE.BYTE_ARRAY_512,
             access=REG_ACCESS.RO,
         ),
         EthernetRegister(
@@ -32,8 +32,8 @@ class EthernetDictionaryV2(DictionaryV2):
             units="",
             subnode=0,
             address=0x00B4,
-            cyclic="CONFIG",
-            dtype=REG_DTYPE.U16,
+            cyclic=RegCyclicType.CONFIG,
+            dtype=REG_DTYPE.BYTE_ARRAY_512,
             access=REG_ACCESS.WO,
         ),
     ]