+++ tests/eoe/test_eoe_dictionary_v3.py
@@ -99,3 +99,46 @@ def test_safety_pdo_not_implemented():
         ethernet_dict.get_safety_rpdo("NOT_EXISTING_UID")
     with pytest.raises(NotImplementedError):
         ethernet_dict.get_safety_tpdo("NOT_EXISTING_UID")
+
+
+@pytest.mark.no_connection
+def test_register_default_values():
+    dictionary_path = join_path(path_resources, dict_eoe_v3)
+    expected_defaults_per_subnode = {
+        0: {
+            "DRV_DIAG_ERROR_LAST_COM": 0,
+            "DRV_AXIS_NUMBER": 1,
+            "CIA301_COMMS_RPDO1_MAP": 1,
+            "CIA301_COMMS_RPDO1_MAP_1": 268451936,
+        },
+        1: {
+            "COMMU_ANGLE_SENSOR": 4,
+        },
+    }
+    ethercat_dict = DictionaryV3(dictionary_path, Interface.EoE)
+    for subnode, registers in ethercat_dict._registers.items():
+        for register in registers.values():
+            assert register.default == expected_defaults_per_subnode[subnode][register.identifier]
+
+
+@pytest.mark.no_connection
+def test_register_description():
+    dictionary_path = join_path(path_resources, dict_eoe_v3)
+    expected_description_per_subnode = {
+        0: {
+            "DRV_DIAG_ERROR_LAST_COM": "Contains the last generated error",
+            "DRV_AXIS_NUMBER": "",
+            "CIA301_COMMS_RPDO1_MAP": "",
+            "CIA301_COMMS_RPDO1_MAP_1": "",
+        },
+        1: {
+            "COMMU_ANGLE_SENSOR": "Indicates the sensor used for angle readings",
+        },
+    }
+    ethercat_dict = DictionaryV3(dictionary_path, Interface.EoE)
+    for subnode, registers in ethercat_dict._registers.items():
+        for register in registers.values():
+            assert (
+                register.description
+                == expected_description_per_subnode[subnode][register.identifier]
+            )